==> Code/retrosynthesis/03.check_whether_metabolite_in_model.ipynb <==
import pandas as pd
from tqdm import tqdm
import re
import os
from util import  *
from rdkit import Chem
tqdm.pandas()
import json
from common import *

#input
yeast_gem_path = '../../Data/model/yeast-GEM.xlsx'
ymdb_corrected_path = '../../Data/database/ymdb/ymdb_correct_smiles.xlsx'
yeast_mat_path = '../../Data/model/yeast-GEM.mat'
chebi_sdf_path = '../../Data/database/ChEBI_complete_3star.sdf'
depr_mnxm_path = '../../Data/database/mnx_chem_depr.tsv'
mnx_chem_prop_path = '../../Data/database/mnx_chem_prop.tsv'
kegg_compound_path = '../../Data/database/kegg_compound.txt'
ymdb_path = '../../Data/database/ymdb/ymdb.csv'

#output
chebi_id = '../../Data/database/chebi_id_smiles.csv'
chebi_second_id = '../../Data/database/chebi_second_id_smiles.csv'
yeast_gem_smiles = '../../Data/yeast_gem_smiles.json'
yeast_gem_final_path = '../../Data/yeast-GEM-final.csv'
ymdb_class_csv = '../../Data/database/ymdb/ymdb_class.csv'
ymdb_output_final = '../../Data/database/ymdb/ymdb_final_result.xlsx'

ymdb = pd.read_csv(ymdb_path)
ymdb_corrected = pd.read_excel(ymdb_corrected_path)
ymdb_class = pd.read_csv(ymdb_class_csv)
ymdb_corrected.head()

for i in range(len(ymdb)):
    if ymdb.at[i, 'ID'] in ymdb_corrected['ID'].values:
        ymdb.at[i, 'SMILES'] = ymdb_corrected[ymdb_corrected['ID'] == ymdb.at[i, 'ID']]['correct_smiles'].values[0]
ymdb.to_csv(ymdb_path, index=False)
for i in range(len(ymdb_class)):
    if ymdb_class.at[i, 'ID'] in ymdb_corrected['ID'].values:
        ymdb_class.at[i, 'SMILES'] = ymdb_corrected[ymdb_corrected['ID'] == ymdb_class.at[i, 'ID']]['correct_smiles'].values[0]
ymdb_class.to_csv(ymdb_class_csv, index=False)

#prepare chebi database
suppl = Chem.SDMolSupplier(chebi_sdf_path)
secondary_chebi = pd.DataFrame(columns=['SMILES','Secondary ChEBI ID'])
chebi = pd.DataFrame(columns=['SMILES','ChEBI ID'])
secondary_smiles_list, secondary_id_list = extract_smiles_and_ids_secondary(suppl)
secondary_chebi['SMILES'] = secondary_smiles_list
secondary_chebi['Secondary ChEBI ID'] = secondary_id_list

secondary_chebi.to_csv(chebi_second_id,index=False)
suppl = Chem.SDMolSupplier(chebi_sdf_path)
smiles_list, id_list = extract_smiles_and_ids(suppl)
chebi['SMILES'] = smiles_list
chebi['ChEBI ID'] = id_list

chebi.to_csv(chebi_id,index=False)

#get smiles from yeast-870 model
metabolite_smiles_dict = get_GEM_all_metabolite_with_smiles(yeast_mat_path)
#save json file
with open(yeast_gem_smiles, 'w') as f:
    json.dump(metabolite_smiles_dict, f,indent=4)

yeast_gem = get_smiles_from_model(yeast_gem_path, yeast_gem_smiles)
yeast_gem.head()

#drop uncorrect smiles
yeast_gem_ = yeast_gem.dropna(subset=['smiles'])
print('Number of metabolites with smiles:', yeast_gem_.shape[0])
yeast_gem = drop_uncorrect_smiles(yeast_gem)
yeast_gem_ = yeast_gem.dropna(subset=['smiles'])
print('Number of metabolites with smiles:', yeast_gem_.shape[0])

yeast_gem['MNXM_ID'] = yeast_gem['MIRIAM'].apply(extract_mnxm_id)
yeast_gem['chebi_id'] = yeast_gem['MIRIAM'].apply(extract_chebi_id)
yeast_gem['kegg_id'] = yeast_gem['MIRIAM'].apply(extract_kegg_id)

# Update metMetaNetXID 
depr_mnxm = pd.read_csv(depr_mnxm_path,sep='\t')
#merged_df = pd.read_csv('../data/metabolites_id.csv')
yeast_gem['metMetaNetXID_new'] = yeast_gem['MNXM_ID'].apply(lambda x: update_metMetaNetXID(x,depr_mnxm['deprecated_ID'].tolist(),depr_mnxm['ID'].tolist()))

#get the smiles from the database and the gem model
database_id = [(chebi_id, 'ChEBI ID','chebi_id'),(mnx_chem_prop_path,'id','MNXM_ID'),(mnx_chem_prop_path,'id','metMetaNetXID_new'),
               (kegg_compound_path, 'KEGG','kegg_id')]
for i in database_id:
    id_csv = i[0]
    extension = os.path.splitext(id_csv)[1]
    if extension == '.csv':
        df = pd.read_csv(id_csv)
    elif extension == '.tsv' or extension == '.txt':
        df = pd.read_csv(id_csv, sep='\t')
    for j in tqdm(range(len(yeast_gem))):
        yeast_gem.loc[j,'smiles'] = update_smiles(yeast_gem.loc[j,'smiles'],df[i[1]].to_list(),df['SMILES'].to_list(),yeast_gem.loc[j,i[2]])
    yeast_gem_ = yeast_gem.dropna(subset=['smiles'])
    print('Number of metabolites with smiles:', yeast_gem_.shape[0])
    yeast_gem = drop_uncorrect_smiles(yeast_gem)
    yeast_gem_ = yeast_gem.dropna(subset=['smiles'])
    print('Number of metabolites with smiles:', yeast_gem_.shape[0])

#get smiles from pubchem 
yeast_gem['smiles'] = yeast_gem.apply(lambda row: add_smiles_pubchem(row['smiles'], row['NAME']), axis=1)
yeast_gem_ = yeast_gem.dropna(subset=['smiles'])
print('Number of metabolites with smiles:', yeast_gem_.shape[0])
yeast_gem = drop_uncorrect_smiles(yeast_gem)
yeast_gem_ = yeast_gem.dropna(subset=['smiles'])
print('Number of metabolites with smiles:', yeast_gem_.shape[0])

#process names of lipids to be in the same format as in ymdb
yeast_gem = yeast_gem.rename(columns={'smiles':'SMILES'})
lipid_df = process_lipid_names(yeast_gem)
sort_fatty_acid_chains(lipid_df, yeast_gem_final_path)

yeast_gem = pd.read_csv(yeast_gem_final_path)
for i in range(len(yeast_gem)):
    if not pd.isnull(yeast_gem.loc[i,'SMILES']):
        if '*' in yeast_gem.loc[i,'SMILES']:
            yeast_gem.loc[i,'SMILES'] = None
yeast_gem['standard_smiles'] = yeast_gem['SMILES'].apply(standardize_smiles)
yeast_gem.to_csv(yeast_gem_final_path,index=False)

#classifying the metabolites in the model
#use the function in ymdb_classify.R, and manually classify the metabolites,save the result in ymdb_class.csv
#Determine whether metablite is in the model
ymdb = pd.read_csv(ymdb_class_csv)
for i in range(len(ymdb)):
    if not pd.isnull(ymdb.loc[i,'SMILES']):
        if '*' in ymdb.loc[i,'SMILES']:
            ymdb.loc[i,'SMILES'] = None
ymdb.loc[ymdb['ID']=='YMDB01162',['NAME']] = 'pa(16:0/16:1)'
ymdb.loc[ymdb['ID']=='YMDB01209',['NAME']] = 'pg(16:0/16:1)'
ymdb['NAME'] = ymdb['NAME'].str.lower()
sort_fatty_acid_chains(ymdb, ymdb_class_csv)
check_metabolite_in_model(yeast_gem_final_path, ymdb_class_csv, ymdb_output_final)

# YMDB metabolites in Yeast8:1139
