==> Code/retrosynthesis/08-5.get_YeastMetaTwin.ipynb <==
import pandas as pd
import numpy as np
import json
import pandas as pd
from cobra import Model, Reaction, Metabolite
from cobra.flux_analysis import flux_variability_analysis
import pickle
from tqdm import tqdm 
from ast import literal_eval
import cobra
import matplotlib.pyplot as plt
import numpy as np
from multiprocessing import Pool
from tqdm import tqdm
from functools import partial 
import matplotlib.pyplot as plt
import numpy as np
from matplotlib.gridspec import GridSpec
from matplotlib.ticker import MaxNLocator, FormatStrFormatter

import gurobipy
gurobipy.setParam('OutputFlag', 0)
import sys
# sys.path.append('../Script/')
from common import *

# input and output

###input
yeast8U_del2_path = '../../Data/model/Yeast-MetaTwin.yml'
ymdb_uptake_file = '../../Data/database/ymdb/ymdb_uptake.csv'

###output
yeast8U_del2_path = '../../Data/model/Yeast-MetaTwin.yml'

# add uptake

model = cobra.io.load_yaml_model(yeast8U_del2_path)

model

ymdb_uptake = pd.read_csv(ymdb_uptake_file)
print(ymdb_uptake.shape)
ymdb_uptake.head(1)

# add YMDB id and YMDB name

metabolites = {}
num = 0
for index, row in ymdb_uptake.iterrows():
    num+=1
    met_id_e = 'snu_'+str(num)
    met_compartment = 'e'
    metabolites[met_id_e] = Metabolite(met_id_e, formula='', name=row['ID'], compartment=met_compartment)

    num+=2
    met_id_c = 'snu_'+str(num)
    met_compartment = 'c'
    metabolites[met_id_c] = Metabolite(met_id_c, formula='', name=row['ID'], compartment=met_compartment)

    reaction = Reaction('rxnu'+str(num))
    reaction.name = 'rxnu'+str(num)
    reaction.subsystem = ''
    reaction.lower_bound = 0.  # This is the default
    reaction.upper_bound = 1000.  # This is the default

    met_e = metabolites.get(met_id_e)
    met_c = metabolites.get(met_id_c)
    reaction.add_metabolites({met_e: -1})
    reaction.add_metabolites({met_c: 1})
    reaction.gene_reaction_rule = ''

    try:
        model.reactions.get_by_id('rxnu'+str(num))
    except:
        model.add_reactions([reaction])    

model

cobra.io.save_yaml_model(model, yeast8U_del2_path)
